Valuation_Company/scripts/investment-news-scraper/requirements.txt
# ================================================================
# 투자 뉴스 스크래핑 - Python 패키지 의존성
# 작성일: 2026-01-25
# ================================================================

# HTTP 요청
requests>=2.31.0

# HTML 파싱
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Supabase 클라이언트
supabase>=2.0.0

# 환경변수 관리
python-dotenv>=1.0.0

# Gemini API
google-generativeai>=0.3.2

# RSS 파싱
feedparser>=6.0.11

# 날짜/시간 처리
python-dateutil>=2.8.0

# PostgreSQL (스키마 적용용)
psycopg2-binary>=2.9.9

# Selenium (동적 사이트 스크래핑 필요 시)
# selenium>=4.15.0
# webdriver-manager>=4.0.0
